feat: add axios and cors
