+++ front-end/package.json
@@ -11,6 +11,7 @@
     "type-check": "vue-tsc --build --force"
   },
   "dependencies": {
+    "axios": "^1.6.8",
     "swiper": "^11.0.7",
     "vue": "^3.4.21",
     "vue-router": "^4.3.0",
